Creating TiltTestBranch

+++ src/com/gpelayo/app/gawalaro/gameObjects/Coordinate.java
@@ -1,10 +1,10 @@
 package com.gpelayo.app.gawalaro.gameObjects;
 
 public class Coordinate {
-	public double x;
-	public double y;
+	public float x;
+	public float y;
 	
-	public Coordinate(double x, double y)
+	public Coordinate(float x, float y)
 	{
 		this.x = x;
 		this.y = y;
